railway-vgg-classification/requirements.txt
# Railway Component Classification - VGG Requirements
# VGG-based classification for railway component condition assessment

# Core ML Frameworks
tensorflow>=2.15.0           # TensorFlow for VGG
keras>=3.0.0                 # Keras for classification

# Computer Vision
opencv-python>=4.8.0         # Image processing
pillow>=10.0.0               # Image handling

# Data Processing
numpy>=1.24.0                # Numerical computing
pandas>=2.0.0                # Data manipulation
scikit-learn>=1.3.0          # ML utilities



# Visualization
matplotlib>=3.7.0            # Plotting
seaborn>=0.12.0              # Statistical visualization

# Utilities
pyyaml>=6.0                  # YAML parsing
requests>=2.31.0             # HTTP requests
tqdm>=4.65.0                 # Progress bars

# Model Export & Deployment
onnx>=1.15.0                 # ONNX export
onnxruntime>=1.16.0          # ONNX runtime
coremltools>=7.0.0           # iOS/macOS deployment

# Optional: Training Monitoring
tensorboard>=2.15.0          # Training visualization
